improved loading indication
