+++ src/charts/util/highlightData.js
@@ -1,4 +1,5 @@
 import { select as d3select } from 'd3';
+import indicateLoading from '../../codebook/util/indicateLoading';
 
 export default function highlightData(chart) {
   const codebook = d3select(
@@ -7,27 +8,29 @@ export default function highlightData(chart) {
     bars = chart.svg.selectAll('.bar-group');
 
   bars.on('click', function(d) {
-    const newIndexes = chart.config.chartType.indexOf('Bars') > -1
-      ? d.values.raw[0].indexes
-      : chart.config.chartType === 'histogramBoxPlot'
-        ? d.values.raw.map(di => di.index)
-        : [];
-    const currentIndexes = codebook.data.highlighted
-        .map(di => di['web-codebook-index']);
-    const removeIndexes = currentIndexes
-        .filter(di => newIndexes.indexOf(di) > -1);
+    indicateLoading(codebook, '.highlightCount', () => {
+        const newIndexes = chart.config.chartType.indexOf('Bars') > -1
+        ? d.values.raw[0].indexes
+        : chart.config.chartType === 'histogramBoxPlot'
+            ? d.values.raw.map(di => di.index)
+            : [];
+        const currentIndexes = codebook.data.highlighted
+            .map(di => di['web-codebook-index']);
+        const removeIndexes = currentIndexes
+            .filter(di => newIndexes.indexOf(di) > -1);
 
-    codebook.data.highlighted = codebook.data.filtered
-          .filter(di => {
-              return removeIndexes.length
-                  ? (currentIndexes.indexOf(di['web-codebook-index']) > -1 && removeIndexes.indexOf(di['web-codebook-index']) === -1)
-                  : (currentIndexes.indexOf(di['web-codebook-index']) > -1 || newIndexes.indexOf(di['web-codebook-index']) > -1);
-    });
+        codebook.data.highlighted = codebook.data.filtered
+            .filter(di => {
+                return removeIndexes.length
+                    ? (currentIndexes.indexOf(di['web-codebook-index']) > -1 && removeIndexes.indexOf(di['web-codebook-index']) === -1)
+                    : (currentIndexes.indexOf(di['web-codebook-index']) > -1 || newIndexes.indexOf(di['web-codebook-index']) > -1);
+        });
 
-    //Display highlighted data in listing & codebook.
-    codebook.data.makeSummary(codebook);
-    codebook.dataListing.init(codebook);
-    codebook.summaryTable.draw(codebook);
-    codebook.controls.updateRowCount(codebook);
+        //Display highlighted data in listing & codebook.
+        codebook.data.makeSummary(codebook);
+        codebook.dataListing.init(codebook);
+        codebook.summaryTable.draw(codebook);
+        codebook.controls.updateRowCount(codebook);
+      });
   });
 }
