+++ src/codebook/summaryTable/draw.js
@@ -1,8 +1,11 @@
 /*------------------------------------------------------------------------------------------------\
   draw/update the summaryTable
 \------------------------------------------------------------------------------------------------*/
+import indicateLoading from '../util/indicateLoading';
 
 export function draw(codebook) {
+    indicateLoading(codebook, '.web-codebook .summaryTable .variable-row .row-title');
+
   //enter/update/exit for variableDivs
   //BIND the newest data
   var varRows = codebook.summaryTable.wrap
